requirements.txt
asgiref             3.8.1
Django              5.0.8
django-cors-headers 4.4.0
djangorestframework 3.15.2
dnspython           2.6.1
pip                 24.0
PyJWT               2.9.0
pymongo             4.8.0
sqlparse            0.5.1
tzdata              2024.1
whitenoise          6.7.0